4 mbuoyb converters (same as those in gdasapp) (#87)
This is a companion to the pr into gdasapp; essentially the same.
https://github.com/NOAA-EMC/GDASApp/pull/1826


4 converters for tropical moorings (rama, pirata, tao/triton) and ndbc
All of these take input from mbouyb
The ndbc contains all the data that is not tropical moorings

The code is basic and repetetive, and will be replaced in the future by
obsbuilder-based converters.
The converters use hard coded lists of wmo ids.
These ids could be read from a file, and the file could be periodically
updated from the erddap database.

---------

Co-authored-by: Edward Givelberg <edwardg@hercules-login-4.hpc.msstate.edu>
Co-authored-by: Edward Givelberg <edwardg@hercules-login-2.hpc.msstate.edu>
Co-authored-by: Edward Givelberg <edwardg@hercules-login-3.hpc.msstate.edu>