Some possible optimizations for AMSR2 (#108)
Sorry, somehow I couldn't figure out how to build on hera, so I didn't
actually try this, but here's the idea for optimization. This is based
on a guess that going to/from string in a loop might be expensive, so
here I:
- moved epoch date creation outside of the loop
- replaced creating a string for the date, and then parsing it in
`DateTime` ctor to passing the year, month etc directly.

Co-authored-by: Anna Shlyaeva <anna.v.shlyaeva@noaa.gov>