Integrate BUFR → IODA conversion for atmospheric observations into obsForge workflow (#146)
This PR addresses the issue of integrating atmospheric observation
BUFR-to-IODA conversion jobs into the obsForge workflow.. See issue:
https://github.com/NOAA-EMC/obsForge/issues/143


Tested all observation types with available script files in SPOC.

The pvlib Python package (required by SSMIS) is temporarily installed in
a personal directory. Since it is not permitted in the production
environment, it will not be included here.

---------

Co-authored-by: xin jin <xin.c.jin@clogin09.cactus.wcoss2.ncep.noaa.gov>