Use IODA-Dump to produce a summary stats file + do not copy Coeff files to COM (#156)
This PR keeps "Coeff" files from being copied back to COM, and also uses
IODA-Dump to produce a summary stats file that contains information on
the number of observations for the Atmos BUFR job. Subsequent PRs should
use this utility to update the blank file to include summary stats for
marine and aero.

---------

Co-authored-by: cory martin <cory.r.martin@clogin08.cactus.wcoss2.ncep.noaa.gov>
Co-authored-by: cory martin <cory.r.martin@clogin04.cactus.wcoss2.ncep.noaa.gov>